All Food Page & Single Page:
This commit implements the 'all food page' which can be navigated to from the navbar. and this page displays all of the food items available in with pagination. This commit also implements the 'single food page' which is the page user gets when the user clicks on the details button on the food item cards on the 'all food page'. The 'single food page' displays the food item with detailed information.