Delete Function on 'My Order' page:
This commit implements the design and functionality to delete any item from the 'My Order' page. When the user clicks on the delete icon in the ui, a fetch requst is sent to delete the specified document from the collection and then the state containing the information about the list in 'my order' page is updated and displayed. In the latest display, the deleted item is removed from the displayed list.