Update & Delete functionality in 'My Added Food Items' page:
This commit implements the update & delete functionality in the 'My added food items' page. Using the update button users can update their item's information in a modal that is displayed when clicked on the update button. And with the delete button user's can easily delete any of their items with a click. and both the result of the delete and update action is displayed immediately on the screen.