Order by number of orders or popularity:
This commit implements the logic of the 'most popular items' section in homepage. This commit stores the number of times a dish has
been ordered in a variable named order. so when the top 6 popular items are retrieved, they are retrieved based on the value of this variable 'order'. And this commit sets up the code to accurately increment that value in the database.