Quantity checking functionality:
This commit implements the checking of available quantity against the user's entry amount, if the user's requested amount exceeds the available amount, then a toast will let the user know that their amount exceeds the available amount, and then the users can decrease their requested quantity. And every time someone orders a dish the available quantity gets calculated and decreased from the database as well. For this I have used the patch api on the server side.